# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_ms-kanban-server_internal_handlers_dto_request.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      permissions:
        additionalProperties:
          additionalProperties:
            type: boolean
          type: object
        type: object
    required:
    - name
    - permissions
    type: object
...
  github_com_ms-kanban-server_internal_handlers_dto_request.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      permissions:
        additionalProperties:
          additionalProperties:
            type: boolean
          type: object
        type: object
    type: object
...
  github_com_ms-kanban-server_internal_handlers_dto_response.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      organization_id:
        type: string
      permissions:
        additionalProperties:
          additionalProperties:
            type: boolean
          type: object
        type: object
      updated_at:
        type: string
    type: object
...
  github_com_ms-kanban-server_internal_pkg_response.Error:
    properties:
      code:
        $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorCode'
      message:
        type: string
      status_code:
        type: integer
    type: object
  github_com_ms-kanban-server_internal_pkg_response.ErrorCode:
    enum:
    - BAD_REQUEST
    - INVALID_STATUS_TRANSITION
    - METHOD_NOT_ALLOWED
    - VALIDATION_ERROR
    - UNAUTHORIZED
    - FORBIDDEN
    - RESOURCE_NOT_FOUND
    - CONFLICT
    - GONE
    - TASK_PERMANENTLY_DELETED
    - BUSINESS_RULE_VIOLATION
    - RATE_LIMIT_EXCEEDED
    - INTERNAL_SERVER_ERROR
    - SERVICE_UNAVAILABLE
    - GATEWAY_TIMEOUT
    type: string
    x-enum-varnames:
    - ErrBadRequest
    - ErrInvalidStatusTransition
    - ErrMethodNotAllowed
    - ErrValidation
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrConflict
    - ErrGone
    - ErrTaskPermanentlyDeleted
    - ErrBusinessRule
    - ErrRateLimitExceeded
    - ErrInternalServerError
    - ErrServiceUnavailable
    - ErrGatewayTimeout
  github_com_ms-kanban-server_internal_pkg_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.Error'
      success:
        type: boolean
    type: object
  github_com_ms-kanban-server_internal_pkg_response.Pagination:
    properties:
      has_next:
        type: boolean
      has_previous:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_ms-kanban-server_internal_pkg_response.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.Pagination'
      status_code:
        type: integer
      success:
        type: boolean
    type: object
info:
  contact:
    email: olivergot26@gmail.com
    name: PMT Team
  description: PMT Backend API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: PMT API
  version: "1.0"
paths:
...
  /organization/roles:
    get:
      consumes:
      - application/json
      description: Lists all custom and system roles available in the organization.
      produces:
      - application/json
      responses:
        "200":
          description: Roles fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.RoleResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Organization Roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Creates a new custom role with custom permissions for the organization.
      parameters:
      - description: Create Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_request.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Custom Role
      tags:
      - Roles
  /organization/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a custom role from the organization.
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Custom Role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Retrieves detail specifications and permission toggles for a given
        role.
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Role Details
      tags:
      - Roles
    patch:
      consumes:
      - application/json
      description: Updates a custom role's name, description, and permissions.
      parameters:
      - description: Role ID
        format: uuid
        in: path
        name: role_id
        required: true
        type: string
      - description: Update Role Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_request.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ms-kanban-server_internal_handlers_dto_response.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ms-kanban-server_internal_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Custom Role
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"